Add initial LABTASKS content
This still needs quite a bit of work,
but this should serve as a good starting point.

In the past versions of lab #3, we've been pretty
explicit about what they should be doing with
Angular, i.e. "make this controller", "make this component".
Thus far, we lack that specificity here.

Issue: #1