Start work on test of `user-list` component.
I've figured out how to mock out the `UserListService` in a simple test of
the `UserListComponent`. Unfortunately I couldn't get it to compile with some
of the nifty filtering stuff in the template, so that's currently commented out,
We obviously want to put that back, however, so it would be awesome
if someone can figure out how to get that to compile and run.

Also, all that `UserListService` does at the moment is return _all_ the users,
which makes the testing kinda dull. It would be nice to extend `UserListService`
so it also makes the call to the server that returns a `User` with a specific ID.
We could then mock that in a more interesting way, and then either test that
some variant of `UserListComponent` displays a single, correct, user, or have
a new single-user component and test it.

I found [this part of the Angular 2 docs](https://angular.io/docs/ts/latest/guide/testing.html#!#component-with-dependency)
useful, and cribbed a lot from their [live examples for `welcome.component.spec.ts`](https://embed.plnkr.co/?show=preview).

Issue #19