Tests *and* production code now work!
There are two major changes here. First, I took the types a lot more seriously and tried to
remove all the use of "cop-out" types like `any` or my little "fake users".

The big discovery, though, was that we should (I think) have been using "subscriptions"
for the component to "listen to" the service (all in `UserListComponent.ngOnInit()`).
This let's the component live in the land of synchronous types (no `Observable<User[]>`)
while the service is still asynchronous. I think that in general you should subscribe to
`Observable`s so you can be notified when there are changes made to them, which can
happen due to things like DB updates.

The filtering still isn't sorted out, but this is *way* closer than we've been before. :-)