The filtering now works!
It turns out that the previous commit essentially fixed the filtering
problems as well, although to be fair I'm not entirely sure why. I think
that the earlier problem had in fact been a type problem (confusion between
`User[]` and `Observable<User[]>`), and fixing that and removing the `async`
from the pipeline totally fixed everything.

Wheee! I think we're ready for class tomorrow. :-)