The filtering now works! (for real)
I someone failed to re-run all the tests when I made the previous commit,
and it turns out that the tests didn't actually pass like I thought they did.
I think, however, that I've finally beaten this into submission, although it's
possible that I've done more than what was strictly necessary; there's a lot
of cobbling together bits of the Internet here.

There were three main changes here:

* First, I created a new `PipeModule` that "holds" our custom filter and allows
people to import it. This has to be added to `AppModule` so everyone has
access to it.
* Then I imported that into the `TestBed` in the `UserListComponent` spec.
* Lastly, I could never get the two-way bindings in the HTML to work, but (a)
that seems to be "not what people do" in A2 and (b) it totally wasn't needed
in this case. Thus I just converted it to a one-way binding where we read from
the input field, and that worked just fine.