Added headless Chrome as an alternative test browser
The default here is to use Chrome, which creates a window.
The headless version doesn't create a window, which may
be necessary for Travis CI.

We used PhantomJS last spring, but recent recommendations
are to use Chrome instead. You also get nicely formatted
and colored test info and error messages in the browser
if you use Chrome.