Get karma tests to run and pass
We updated the (quite limited) tests using what appears to
be more "Angular 4" style and got those tests to run and
pass.

One weird thing that took a while to figure out was
setting `APP_BASE_REF` in `app.module.ts`, which was
crucial to getting the tests to even run.

We also substantially rewrote `app.component.spec.ts` to
follow the style in the Angular 4 testing examples. It
took us a while to figure out that we needed to import
the `AppModule` to make all the dependent components
available to the test runner.