Set up single run tests and watch tests
This gives us a new `gradle` task called `testWatch` that runs the
test framework in "watch" mode (so any changes are seen and the tests
are re-run). The old `runClientTests` option is still there, but only
runs the tests once and then stops. This means that `gradle test` now
runs both the Java and JS tests without getting hung because the JS
tests are in "watch" mode.