Add preliminary e2e test and add gradle task for e2e
Test 1 checks and highlights that the title "User Name" is there
Test 2 is a little weird, the server must be running first for users to be
on seen. So this test will fail if its not running but I was just playing around
Test 3 "will" check that after typing in Lynn it will filter and the sentence with
Lynn will be appear ... but first filter needs to be implemented.