Improve error handling in `UserListComponent`
If there was an error retrieving a list of users in `UserListComponent`
we just logged it and gave the user no useful feedback. This at least
puts some info on the page saying there was a problem. I kinda think
this should actually be a dialog rather than just text on the page,
but I couldn't figure that out quickly so this will do for now.