Add multiple hint support to the play component
This changes it so that the card is stored in the cardstate. This means that there is only one array for the slider and the card states are no longer a seperate thing. When the component first gets the cards from the deck it maps them into card states, randomizing the hints at the same time. The component is also set to only get the cards once instead of continuing to update the array when it changes. This prevents the randomization from changing and prevents problems if a hint were to be removed while someone is playing the deck.