Added Card History
Added card history which shows who made the card, when they made it, who last edited the card, and when it was last edited. These only show when they are relevant, for instance who made the card and when wont show if the card was made before the feature was added.